mcp_server/requirements.txt
fastapi>=0.100.0,<0.111.0 # Pinning minor version for stability
uvicorn[standard]>=0.20.0,<0.28.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
requests>=2.28.0 # Often a dependency of LLM libraries, good to have explicit
# LLM Libraries - Install based on which ones you need
google-generativeai>=0.4.0,<0.6.0
openai>=1.0.0,<2.0.0
anthropic>=0.20.0,<0.26.0
